added a user entity

+++ src/tasks/tasks.controller.ts
@@ -1,20 +1,6 @@
-import {
-  Body,
-  Controller,
-  Delete,
-  Get,
-  Patch,
-  Post,
-  UseGuards,
-  UsePipes,
-} from '@nestjs/common';
+import { Body, Controller, Delete, Get, Patch, Post, Req, UseGuards, UsePipes } from '@nestjs/common';
 import { TasksService } from './tasks.service';
-import {
-  ICreateTask,
-  IDeleteTask,
-  IGetTask,
-  IUpdateTask,
-} from './tasks.interface';
+import { ICreateTask, IDeleteTask, IGetTask, IUpdateTask } from './tasks.interface';
 import { JoiValidationPipe } from 'src/common/pipes/joi.pipe';
 import {
   createTaskValidations,
@@ -34,32 +20,36 @@ export class TasksController {
   @Get('get')
   @UseGuards(JWTAuthGuard)
   @UsePipes(new JoiValidationPipe(getTasksValidations))
-  async get(@Body() getTaskDto: IGetTask) {
-    const task = await this.taskService.get(getTaskDto);
+  async get(@Body() getTaskDto: IGetTask, @Req() req: Request) {
+    const userId = req.headers['userId'];
+    const task = await this.taskService.get({ ...getTaskDto, userId });
     return task;
   }
 
   @Post('/create')
   @UseGuards(JWTAuthGuard)
   @UsePipes(new JoiValidationPipe(createTaskValidations))
-  async create(@Body() createTaskDto: ICreateTask) {
-    const task = await this.taskService.create(createTaskDto);
+  async create(@Body() createTaskDto: ICreateTask, @Req() req: Request) {
+    const userId = req.headers['userId'];
+    const task = await this.taskService.create({ ...createTaskDto, userId });
     return task;
   }
 
   @Patch('/update')
   @UseGuards(JWTAuthGuard)
   @UsePipes(new JoiValidationPipe(updateTaskValidations))
-  async update(@Body() updateTaskDto: IUpdateTask) {
-    const task = await this.taskService.update(updateTaskDto);
+  async update(@Body() updateTaskDto: IUpdateTask, @Req() req: Request) {
+    const userId = req.headers['userId'];
+    const task = await this.taskService.update({ ...updateTaskDto, userId });
     return task;
   }
 
   @Delete('/delete')
   @UseGuards(JWTAuthGuard)
   @UsePipes(new JoiValidationPipe(deleteTaskValidations))
-  async delete(@Body() deleteTaskDto: IDeleteTask) {
-    const task = await this.taskService.delete(deleteTaskDto);
+  async delete(@Body() deleteTaskDto: IDeleteTask, @Req() req: Request) {
+    const userId = req.headers['userId'];
+    const task = await this.taskService.delete({ ...deleteTaskDto, userId });
     return task;
   }
 }
